requirements.txt
restate-sdk==0.5.1
hypercorn
minio
langchain
langchain_community
langchain-ollama
langchain-qdrant
pypdf
pandas
